// src/shared/enums/enums.ts
export enum ButtonThemes {
  Primary = 'primary',
  Outline = 'outline',
  Add = 'add',
}

export enum InputThemes {
  Gray = 'gray',
  White = 'white',
}

export enum Steps {
  One = 1,
  Two = 2,
  Three = 3,
}

export enum RouteEndpoints {
  Home = '/',
  Form = 'create',
}

export enum FormInputs {
  Phone = 'phone',
  Email = 'email',
  Nickname = 'nickname',
  Name = 'name',
  Surname = 'surname',
  Sex = 'sex',
  Advantages = 'advantages',
  AdvantageItem = 'advantage-item',
  CheckboxGroup = 'checkbox-group',
  RadioGroup = 'radio-group',
  About = 'about',
}